migrate jackson, rewrite big decimal logic so older versions than 2.12 can be supported
