+++ build.gradle
@@ -36,9 +36,9 @@ repositories {
 }
 
 def jsonProviders = [
-        [id   : 'jackson', additionalVersions: [],
+        [id   : 'jackson', additionalVersions: ['2.2.0'],
          group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.18.0'],
-        [id   : 'gson', additionalVersions: ['2.3', '2.6.2', '2.8.9', '2.9.1', '2.10.1'],
+        [id   : 'gson', additionalVersions: ['2.3'],
          group: 'com.google.code.gson', name: 'gson', version: '2.11.0'],
         [id   : 'orgJson', additionalVersions: [],
          group: 'org.json', name: 'json', version: '20240303'],
